# Speed up inflate_fast by up to 1.5x on x86_64.