@nigeltao 2017-10-07
I'll add some detail.

There are two similar but separable techniques that speed up zlib's inflate_fast function (decompression). Let's call those techniques T1 and T2.

T1 is that, when filling the bit buffer for Huffman decoding, that buffer can be filled 8 bytes at a time, instead of 1 byte at a time. This technique requires 64-bit unaligned and also little-endian loads. On x86_64, this can give up to a 1.2x speed up.

T2 is that, when performing "copy LENGTH bytes from previously decompressed output, DIST bytes ago", that copy of LENGTH bytes can often be done 8 bytes at a time, instead of 1 byte at a time. Any given iteration can copy too many bytes, but subsequent iterations will fix the overrun. The final iteration might not be fixed up this way, but the inflate_fast function returns the number of bytes written, and that number can correct for any final overrun. This technique requires 64-bit unaligned loads and stores. On x86_64, this can give up to a 1.3x speed up.

In inffast.c, T1 code is #ifdef ZLIB_INTERNAL_HAVE_64_BIT_UNALIGNED_LOADS_LITTLE_ENDIAN. T2 code is #ifdef ZLIB_INTERNAL_HAVE_64_BIT_UNALIGNED_LOADS. Note that the first macro ends in "LITTLE_ENDIAN" and the second does not.

Together, this change can give up to a 1.5x speed up (on x86_64).

T1 should cause no change in the decoded output whatsoever. T2 can produce different output, in the sense that the zlib API is essentially "decode into this buffer (i.e. char* buf, int len as args) and zlib will return an int (call it n) that is the number of bytes decoded into that buffer". In Go-like slice notation, you would expect the caller to be interested in buf[:n], the decompressed data, and not buf[n:], 'scratch space'. T2 might write different bytes into the scratch space buf[n:len], but the code was careful (via a larger INFLATE_FAST_MIN_LEFT threshold) to never write a buffer overflow, into buf[len:].

I had convinced myself that that scratch space was safe to overrun. As I said, I couldn't think of why callers would care about buf[n:]. The punchline is that T2 as is does have a correctness bug.

Zlib-the-library actually speaks three related formats (RFCs 1950, 1951 and 1952). Gzip and zlib-the-format are small header/footers wrapped around raw flate. Many users of the library use one of those two wrapped format. For example, HTTP's Content-Encoding speaks gzip. The PNG image format speaks zlib. On the other hand, ZIP uses raw flate.

Zlib-the-library also provides two similar but different APIs. Decoding the two wrapped formats (gzip and zlib) involves calling three functions: inflateInit2, inflate, inflateEnd, all implemented in inflate.c.

There is a different set of functions to call for the raw format: inflateBackInit, inflateBack, inflateBackEnd, all implemented in infback.c. I guess that "Back" is short for "callback", because the API takes a callback. The inflate.c code could could conceivably handle raw flate as well, with a small patch, but zlib-the-library is what it is, possibly for historical reasons.

In any case, the inflate.c and the infback.c code both call inflate_fast (in inffast.c), but with different buffer allocation patterns. inflate.c malloc's its own history buffer, also known as the 'window', as a private field. infback.c takes a window buffer (pointer and length) as an argument, and uses that buffer both as the history buffer and as the destination buffer for inflate_fast: the destination is the "buf" in "buf[n:]" that I described some paragraphs above.

Thus, buf[n:] isn't always just 'scratch space', and having T2 overwrite into that buffer in one iteration can actually (incorrectly) change the decoded bytes in a subsequent iteration when executing a flate distance-length back-copy from the window. This is still pretty rare, but not impossible.

Thus, I need to walk back T2's enthusiasm for 8 byte copies, and ensure that the entirety of buf[:len] is unchanged, not just buf[:n]. As somebody wiser than me could have forseen, in any case, zlib is old enough and well-used enough that Hyrum's law (http://www.hyrumslaw.com/) possibly applies.

The T2 technique is not fundamentally flawed, but it needs more thought. Some of the performance gain could be recovered if we introduced additional API (or deleted some API in a hypothetical backwards-incompatible "zlib version 2.0") that disallowed aliasing the window and destination buffer, but that is a much more complicated change, out of scope of this pull request.